Fixed some codes and added ID because it cant update

+++ SAOProject/SettingsForm.vb
@@ -5,11 +5,31 @@ Public Class SettingsForm
 
     Dim question As DialogResult
 
+    Public identifier_rankpi As String
+    Public identifier_ranorg As String
+    Public identifier_ranloc As String
+    Public random As System.Random = New System.Random
+
 
 
     Private Sub SettingsForm_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         load_organizations()
         load_location()
+        auto_generate_id()
+    End Sub
+    Public Sub auto_generate_id()
+        identifier_rankpi = Now.ToString("mmddyyy" + "-")
+        identifier_rankpi = identifier_rankpi + random.Next(1, 100000).ToString
+
+        identifier_ranloc = Now.ToString("mmddyyy" + "-")
+        identifier_ranloc = identifier_ranloc + random.Next(1, 100000).ToString
+
+        identifier_ranorg = Now.ToString("mmddyyy" + "-")
+        identifier_ranorg = identifier_ranorg + random.Next(1, 100000).ToString
+
+        ran_kpi.Text = identifier_rankpi
+        ran_loc.Text = identifier_ranloc
+        ran_org.Text = identifier_ranorg
     End Sub
 
     Public Sub load_organizations()
@@ -22,7 +42,7 @@ Public Class SettingsForm
             Dim dbdataset As New DataTable
 
             MysqlConn.Open()
-            query = "SELECT school as 'School/Organization' FROM tbl_organizations_school"
+            query = "SELECT orgid as 'ID',school as 'School/Organization' FROM tbl_organizations_school"
             Command = New MySqlCommand(query, MysqlConn)
             SDA.SelectCommand = Command
             SDA.Fill(dbdataset)
@@ -36,6 +56,9 @@ Public Class SettingsForm
             MessageBox.Show(ex.Message)
         Finally
             MysqlConn.Dispose()
+
+            Dim columndrugname = mg_organizations.Columns(0)
+            columndrugname.Width = 100
         End Try
     End Sub
 
@@ -49,7 +72,7 @@ Public Class SettingsForm
             Dim dbdataset As New DataTable
 
             MysqlConn.Open()
-            query = "SELECT location as 'Location/s' FROM tbl_locations"
+            query = "SELECT loc_id as 'ID',location as 'Location/s' FROM tbl_locations"
             Command = New MySqlCommand(query, MysqlConn)
             SDA.SelectCommand = Command
             SDA.Fill(dbdataset)
@@ -63,6 +86,9 @@ Public Class SettingsForm
             MessageBox.Show(ex.Message)
         Finally
             MysqlConn.Dispose()
+
+            Dim columndrugname = mg_locations.Columns(0)
+            columndrugname.Width = 100
         End Try
     End Sub
 
@@ -118,10 +144,11 @@ Public Class SettingsForm
                     Else
                         MysqlConn.Close()
                         MysqlConn.Open()
-                        query = "INSERT INTO tbl_organizations_school VALUES (@school)"
-                            Command = New MySqlCommand(query, MysqlConn)
-                            Command.Parameters.AddWithValue("school", tb_organization.Text)
-                            reader = Command.ExecuteReader
+                        query = "INSERT INTO tbl_organizations_school VALUES (@school,@orgid)"
+                        Command = New MySqlCommand(query, MysqlConn)
+                        Command.Parameters.AddWithValue("school", tb_organization.Text)
+                        Command.Parameters.AddWithValue("orgid", ran_org.Text)
+                        reader = Command.ExecuteReader
 
                             MetroMessageBox.Show(Me, "Successfully Saved!", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Information)
                         End If
@@ -137,6 +164,7 @@ Public Class SettingsForm
         Finally
             MysqlConn.Dispose()
             load_organizations()
+            auto_generate_id()
         End Try
 
     End Sub
@@ -150,7 +178,6 @@ Public Class SettingsForm
                 MysqlConn.Close()
             End If
 
-
             If (tb_organization.Text = "") Then
                 MetroMessageBox.Show(Me, "Please fill the fields", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Warning)
             Else
@@ -161,15 +188,16 @@ Public Class SettingsForm
 
 
                     MysqlConn.Open()
-                    query = "UPDATE tbl_organizations_school SET school=@school"
+                    query = "UPDATE tbl_organizations_school SET school=@school WHERE orgid=@orgid "
                     Command = New MySqlCommand(query, MysqlConn)
-                            Command.Parameters.AddWithValue("school", tb_organization.Text)
-                            reader = Command.ExecuteReader
+                    Command.Parameters.AddWithValue("school", tb_organization.Text)
+                    Command.Parameters.AddWithValue("orgid", ran_org.Text)
+                    reader = Command.ExecuteReader
 
                     MetroMessageBox.Show(Me, "Successfully Updated!", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Information)
                 End If
 
-                    End If
+            End If
 
 
         Catch ex As MySqlException
@@ -177,11 +205,9 @@ Public Class SettingsForm
 
         Finally
             MysqlConn.Dispose()
+            auto_generate_id()
             load_organizations()
         End Try
-
-
-
     End Sub
 
     Private Sub mbtn_delete_Click(sender As Object, e As EventArgs) Handles mbtn_delete.Click
@@ -204,41 +230,39 @@ Public Class SettingsForm
 
 
                     MysqlConn.Open()
-                    query = "DELETE FROM tbl_organizations_school WHERE school=@school"
+                    query = "DELETE FROM tbl_organizations_school WHERE school=@school AND orgid=@orgid"
                     Command = New MySqlCommand(query, MysqlConn)
                     Command.Parameters.AddWithValue("school", tb_organization.Text)
+                    Command.Parameters.AddWithValue("orgid", ran_org.Text)
                     reader = Command.ExecuteReader
 
                     MetroMessageBox.Show(Me, "Successfully Deleted!", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Information)
                 End If
-
             End If
-
-
         Catch ex As MySqlException
             MessageBox.Show(ex.Message)
 
         Finally
             MysqlConn.Dispose()
             tb_organization.Clear()
+            auto_generate_id()
             load_organizations()
         End Try
     End Sub
     Private Sub mbtn_clearorg_Click(sender As Object, e As EventArgs) Handles mbtn_clearorg.Click
         tb_organization.Clear()
-
+        auto_generate_id()
     End Sub
-
-    Private Sub mg_organizations_CellContentClick(sender As Object, e As DataGridViewCellEventArgs) Handles mg_organizations.CellContentClick
-
+    Private Sub mg_organizations_CellClick(sender As Object, e As DataGridViewCellEventArgs) Handles mg_organizations.CellClick
         If e.RowIndex >= 0 Then
             Dim row As DataGridViewRow
             row = Me.mg_organizations.Rows(e.RowIndex)
 
             tb_organization.Text = row.Cells("School/Organization").Value.ToString
-
+            ran_org.Text = row.Cells("ID").Value.ToString
         End If
     End Sub
+
 
     Private Sub mbtn_locationsave_Click(sender As Object, e As EventArgs) Handles mbtn_locationsave.Click
         Try
@@ -272,19 +296,16 @@ Public Class SettingsForm
                     MetroMessageBox.Show(Me, "The Location " & tb_location.Text & " is already registered.", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Warning)
 
                 Else
-                    If (tb_location.Text = "") Then
-                        MetroMessageBox.Show(Me, "Please fill all fields", "Student Affairs Office Consolidated Calendar", MessageBoxButtons.OK, MessageBoxIcon.Information)
-                    Else
-                        MysqlConn.Close()
+
+                    MysqlConn.Close()
                         MysqlConn.Open()
-                        query = "INSERT INTO tbl_locations VALUES (@location)"
-                        Command = New MySqlCommand(query, MysqlConn)
-                        Command.Parameters.AddWithValue("location", tb_location.Text)
-                        reader = Command.ExecuteReader
+                    query = "INSERT INTO tbl_locations VALUES (@location,@loc_id)"
+                    Command = New MySqlCommand(query, MysqlConn)
+                    Command.Parameters.AddWithValue("location", tb_location.Text)
+                    Command.Parameters.AddWithValue("loc_id", ran_loc.Text)
+                    reader = Command.ExecuteReader
 
                         MetroMessageBox.Show(Me, "Successfully Saved!", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Information)
-                    End If
-
 
                 End If
 
@@ -295,8 +316,104 @@ Public Class SettingsForm
         Finally
             MysqlConn.Dispose()
             load_location()
+            auto_generate_id()
         End Try
     End Sub
 
+    Private Sub mbtn_updatelocations_Click(sender As Object, e As EventArgs) Handles mbtn_updatelocations.Click
+        Try
+            MysqlConn = New MySqlConnection
+            MysqlConn.ConnectionString = connstring
 
+            If MysqlConn.State = ConnectionState.Open Then
+                MysqlConn.Close()
+            End If
+
+
+            If (tb_location.Text = "") Then
+                MetroMessageBox.Show(Me, "Please fill the fields", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Warning)
+            Else
+
+                question = MetroMessageBox.Show(Me, "Are you sure you want to update this?", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.YesNo, MessageBoxIcon.Question)
+
+                If (question = DialogResult.Yes) Then
+
+
+                    MysqlConn.Open()
+                    query = "UPDATE tbl_locations SET Location=@location WHERE loc_id=@loc_id"
+                    Command = New MySqlCommand(query, MysqlConn)
+                    Command.Parameters.AddWithValue("location", tb_location.Text)
+                    Command.Parameters.AddWithValue("loc_id", ran_loc.Text)
+                    reader = Command.ExecuteReader
+
+                    MetroMessageBox.Show(Me, "Successfully Updated!", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Information)
+                End If
+
+            End If
+
+
+        Catch ex As MySqlException
+            MessageBox.Show(ex.Message)
+
+        Finally
+            MysqlConn.Dispose()
+            load_location()
+            auto_generate_id()
+        End Try
+    End Sub
+
+    Private Sub mbtn_deletelocations_Click(sender As Object, e As EventArgs) Handles mbtn_deletelocations.Click
+        Try
+            MysqlConn = New MySqlConnection
+            MysqlConn.ConnectionString = connstring
+
+            If MysqlConn.State = ConnectionState.Open Then
+                MysqlConn.Close()
+            End If
+
+
+            If (tb_location.Text = "") Then
+                MetroMessageBox.Show(Me, "Please choose from the table.", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Warning)
+            Else
+
+                question = MetroMessageBox.Show(Me, "Are you sure you want to delete this?", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.YesNo, MessageBoxIcon.Question)
+
+                If (question = DialogResult.Yes) Then
+
+
+                    MysqlConn.Open()
+                    query = "DELETE FROM tbl_locations WHERE Location=@location AND loc_id=@loc_id"
+                    Command = New MySqlCommand(query, MysqlConn)
+                    Command.Parameters.AddWithValue("location", tb_location.Text)
+                    Command.Parameters.AddWithValue("loc_id", ran_loc.Text)
+                    reader = Command.ExecuteReader
+
+                    MetroMessageBox.Show(Me, "Successfully Deleted!", "Students Affairs Office Consolidated Calendar ", MessageBoxButtons.OK, MessageBoxIcon.Information)
+                End If
+            End If
+        Catch ex As MySqlException
+            MessageBox.Show(ex.Message)
+
+        Finally
+            MysqlConn.Dispose()
+            tb_location.Clear()
+            auto_generate_id()
+            load_location()
+        End Try
+    End Sub
+
+    Private Sub mbtn_clearloc_Click(sender As Object, e As EventArgs) Handles mbtn_clearloc.Click
+        auto_generate_id()
+        tb_location.Clear()
+    End Sub
+
+    Private Sub mg_locations_CellClick(sender As Object, e As DataGridViewCellEventArgs) Handles mg_locations.CellClick
+        If e.RowIndex >= 0 Then
+            Dim row As DataGridViewRow
+            row = Me.mg_locations.Rows(e.RowIndex)
+
+            tb_location.Text = row.Cells("Location/s").Value.ToString
+            ran_loc.Text = row.Cells("ID").Value.ToString
+        End If
+    End Sub
 End Class
